move CSS functions into own module
